:truck: 패키지 구조 변경
- 도메인 패키지와 dto 패키지 분리
