+++ noriter/src/main/java/com/codewarts/noriter/article/controller/StudyController.java
@@ -1,9 +1,9 @@
 package com.codewarts.noriter.article.controller;
 
-import com.codewarts.noriter.article.domain.dto.study.StudyDetailResponse;
-import com.codewarts.noriter.article.domain.dto.study.StudyEditRequest;
-import com.codewarts.noriter.article.domain.dto.study.StudyListResponse;
-import com.codewarts.noriter.article.domain.dto.study.StudyPostRequest;
+import com.codewarts.noriter.article.dto.study.StudyDetailResponse;
+import com.codewarts.noriter.article.dto.study.StudyEditRequest;
+import com.codewarts.noriter.article.dto.study.StudyListResponse;
+import com.codewarts.noriter.article.dto.study.StudyPostRequest;
 import com.codewarts.noriter.article.domain.type.StatusType;
 import com.codewarts.noriter.article.service.StudyService;
 import com.codewarts.noriter.exception.GlobalNoriterException;
