+++ noriter/src/main/java/com/codewarts/noriter/article/dto/free/FreeListResponse.java
@@ -1,4 +1,4 @@
-package com.codewarts.noriter.article.domain.dto.free;
+package com.codewarts.noriter.article.dto.free;
 
 import com.codewarts.noriter.article.domain.Article;
 import com.codewarts.noriter.article.domain.Hashtag;
