+++ noriter/src/main/java/com/codewarts/noriter/article/dto/question/QuestionListResponse.java
@@ -1,10 +1,8 @@
-package com.codewarts.noriter.article.domain.dto.question;
+package com.codewarts.noriter.article.dto.question;
 
 import com.codewarts.noriter.article.domain.Hashtag;
 import com.codewarts.noriter.article.domain.Question;
 import com.codewarts.noriter.article.domain.type.StatusType;
-import com.codewarts.noriter.comment.domain.dto.CommentResponse;
-import com.codewarts.noriter.member.domain.dto.WriterInfoResponse;
 import com.fasterxml.jackson.annotation.JsonFormat;
 import java.time.LocalDateTime;
 import java.util.List;
@@ -12,12 +10,12 @@ import java.util.stream.Collectors;
 import lombok.Getter;
 
 @Getter
-public class QuestionDetailResponse {
+public class QuestionListResponse {
 
     private final Long id;
     private final String title;
     private final String content;
-    private final WriterInfoResponse writer;
+    private final String writerNickname;
     private final boolean sameWriter;
     private final List<String> hashtags;
     @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
@@ -26,29 +24,21 @@ public class QuestionDetailResponse {
     private final LocalDateTime editedTime;
     private final boolean wish;
     private final int wishCount;
-    private final List<CommentResponse> comment;
+    private final int commentCount;
     private final StatusType status;
 
-    public QuestionDetailResponse(Question question, boolean sameWriter, boolean wish) {
+    public QuestionListResponse(Question question, boolean sameWriter, boolean wish) {
         this.id = question.getId();
         this.title = question.getTitle();
         this.content = question.getContent();
-        this.writer = WriterInfoResponse.from(question.getWriter());
+        this.writerNickname = question.getWriter().getNickname();
         this.sameWriter = sameWriter;
-        this.hashtags = question.getHashtags().stream()
-            .map(Hashtag::getContent)
-            .collect(Collectors.toList());
+        this.hashtags = question.getHashtags().stream().map(Hashtag::getContent).collect(Collectors.toList());
         this.writtenTime = question.getWrittenTime();
         this.editedTime = question.getEditedTime();
         this.wish = wish;
         this.wishCount = question.getWishList().size();
-        this.comment = question.getComments().stream()
-            .map(CommentResponse::new)
-            .collect(Collectors.toList());
+        this.commentCount = question.getComments().size();
         this.status = question.getStatus();
-    }
-
-    public static QuestionDetailResponse from(Question question, boolean sameWriter, boolean wish) {
-        return new QuestionDetailResponse(question, sameWriter, wish);
     }
 }
