+++ noriter/src/main/java/com/codewarts/noriter/article/service/QuestionService.java
@@ -1,10 +1,10 @@
 package com.codewarts.noriter.article.service;
 
 import com.codewarts.noriter.article.domain.Question;
-import com.codewarts.noriter.article.domain.dto.question.QuestionDetailResponse;
-import com.codewarts.noriter.article.domain.dto.question.QuestionListResponse;
-import com.codewarts.noriter.article.domain.dto.question.QuestionPostRequest;
-import com.codewarts.noriter.article.domain.dto.question.QuestionUpdateRequest;
+import com.codewarts.noriter.article.dto.question.QuestionDetailResponse;
+import com.codewarts.noriter.article.dto.question.QuestionListResponse;
+import com.codewarts.noriter.article.dto.question.QuestionPostRequest;
+import com.codewarts.noriter.article.dto.question.QuestionUpdateRequest;
 import com.codewarts.noriter.article.domain.type.StatusType;
 import com.codewarts.noriter.article.repository.ArticleRepository;
 import com.codewarts.noriter.article.repository.QuestionRepository;
